# docs/api_contract.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Equity Lab",
    "description": "Institutional Indian Equity Research, Options Arbitrage, and Return Probability Engine",
    "version": "0.0.9"
  },
  "paths": {
...
    "/api/v1/regime": {
      "get": {
        "tags": [
          "Market Data"
        ],
        "summary": "Fetch Market Regime",
        "description": "Evaluates live India VIX and Nifty volatility regime.",
        "operationId": "fetch_market_regime_api_v1_regime_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketRegimeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/technical/report/{symbol}": {
      "get": {
        "tags": [
          "Institutional Technical Framework"
        ],
        "summary": "Get Technical Report",
        "description": "Retrieves full standardized Machine-Readable Technical Report for a given ticker symbol.",
        "operationId": "get_technical_report_api_v1_technical_report__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Technical Report Api V1 Technical Report  Symbol  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/technical/regime": {
      "get": {
        "tags": [
          "Institutional Technical Framework"
        ],
        "summary": "Get Market Regime",
        "description": "Retrieves current market regime classification for Indian equities.",
        "operationId": "get_market_regime_api_v1_technical_regime_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Market Regime Api V1 Technical Regime Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/technical/screener": {
      "get": {
        "tags": [
          "Institutional Technical Framework"
        ],
        "summary": "Run Screener",
        "description": "Runs 3-tier technical universe screener and returns top setup candidates.",
        "operationId": "run_screener_api_v1_technical_screener_get",
        "parameters": [
          {
            "name": "min_tss_score",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 100.0,
              "minimum": 0.0,
              "default": 65.0,
              "title": "Min Tss Score"
            }
          },
          {
            "name": "setup_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Setup Filter"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Run Screener Api V1 Technical Screener Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/technical/probability/{symbol}": {
      "get": {
        "tags": [
          "Institutional Technical Framework"
        ],
        "summary": "Get Probability Ladder",
        "description": "Retrieves empirical probability ladder and MAE/MFE path statistics.",
        "operationId": "get_probability_ladder_api_v1_technical_probability__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Probability Ladder Api V1 Technical Probability  Symbol  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/technical/trade_manager/{symbol}": {
      "get": {
        "tags": [
          "Institutional Technical Framework"
        ],
        "summary": "Get Trade Management",
        "description": "Evaluates active in-position trade management rules for a live position.",
        "operationId": "get_trade_management_api_v1_technical_trade_manager__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "entry_price",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "exclusiveMinimum": 0.0,
              "default": 500.0,
              "title": "Entry Price"
            }
          },
          {
            "name": "current_price",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "exclusiveMinimum": 0.0,
              "default": 525.0,
              "title": "Current Price"
            }
          },
          {
            "name": "highest_close",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "exclusiveMinimum": 0.0,
              "default": 530.0,
              "title": "Highest Close"
            }
          },
          {
            "name": "stop_price",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "exclusiveMinimum": 0.0,
              "default": 475.0,
              "title": "Stop Price"
            }
          },
          {
            "name": "days_in_trade",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 6,
              "title": "Days In Trade"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Trade Management Api V1 Technical Trade Manager  Symbol  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/technical/surveillance/{symbol}": {
      "get": {
        "tags": [
          "Institutional Technical Framework"
        ],
        "summary": "Get Surveillance Gate",
        "description": "Retrieves ASM/GSM/T2T regulatory surveillance status and roundtrip trade cost breakdown.",
        "operationId": "get_surveillance_gate_api_v1_technical_surveillance__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Surveillance Gate Api V1 Technical Surveillance  Symbol  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/technical/short-term-prediction/{symbol}": {
      "get": {
        "tags": [
          "Institutional Technical Framework"
        ],
        "summary": "Get Short Term Prediction",
        "description": "Retrieves high-precision short-term multi-horizon prediction, AVWAPs, Stoch RSI exhaustion, and Conformal Cones.",
        "operationId": "get_short_term_prediction_api_v1_technical_short_term_prediction__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Short Term Prediction Api V1 Technical Short Term Prediction  Symbol  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/technical/chart/reconcile": {
      "post": {
        "tags": [
          "Institutional Technical Framework"
        ],
        "summary": "Reconcile Chart Vision",
        "description": "Reconciles visual chart features (pixels/vision extraction) against authoritative numerical OHLCV exchange truth.",
        "operationId": "reconcile_chart_vision_api_v1_technical_chart_reconcile_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Reconcile Chart Vision Api V1 Technical Chart Reconcile Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MarketRegimeResponse": {
        "properties": {
          "vix_level": {
            "type": "number",
            "title": "Vix Level"
          },
          "regime": {
            "type": "string",
            "title": "Regime"
          },
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "a2_suitability": {
            "type": "string",
            "title": "A2 Suitability"
          },
          "observation": {
            "type": "string",
            "title": "Observation"
          },
          "nifty_spot": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nifty Spot"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaHeader"
          }
        },
        "type": "object",
        "required": [
          "vix_level",
          "regime",
          "score",
          "a2_suitability",
          "observation",
          "meta"
        ],
        "title": "MarketRegimeResponse"
      },
      "MetaHeader": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source",
            "description": "Name of the underlying market data feed/provider"
          },
          "as_of": {
            "type": "string",
            "title": "As Of",
            "description": "Timestamp in IST (Asia/Kolkata) when market data was generated"
          },
          "retrieved_at": {
            "type": "string",
            "title": "Retrieved At",
            "description": "Timestamp in IST when backend fetched the data"
          },
          "market_data_type": {
            "type": "string",
            "title": "Market Data Type",
            "description": "'realtime', 'delayed', or 'end_of_day'"
          },
          "stale": {
            "type": "boolean",
            "title": "Stale",
            "description": "True if data is older than expected TTL",
            "default": false
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations",
            "description": "Explicit limits, disclaimers, or notices"
          }
        },
        "type": "object",
        "required": [
          "source",
          "as_of",
          "retrieved_at",
          "market_data_type"
        ],
        "title": "MetaHeader"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}